fix confusing elided lifetimes
