+++ src/stable/lib/canisters/nns_icp_ledger/idl/README.md
@@ -7,12 +7,3 @@
 2. Find the Candid on that page and copy it
 3. Paste the Candid into `index.did`
 4. From this directory run `npx azle generate index.did > index.ts`
-5. Add the following to the bottom of `index.ts`:
-
-```typescript
-export {
-    binaryAddressFromAddress,
-    binaryAddressFromPrincipal,
-    hexAddressFromPrincipal
-} from './address';
-```
